Try to fix build error by splitting Messaging file
