+++ src/frontend/containers/AppToolbar/SecondaryCommands.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react-lite';
 
 import { IconSet, KeyCombo } from 'widgets';
 import { MenuButton, MenuItem } from 'widgets/menus';
-import { RendererMessenger } from 'src/Messaging';
+import { RendererMessenger } from 'src/ipc/RenderMessenger';
 import { useStore } from 'src/frontend/contexts/StoreContext';
 
 const SecondaryCommands = observer(() => {
