+++ src/frontend/containers/WindowTitlebar.tsx
@@ -1,6 +1,7 @@
 import { observer } from 'mobx-react-lite';
 import React, { useEffect, useState } from 'react';
-import { RendererMessenger, WindowSystemButtonPress } from 'src/Messaging';
+import { WindowSystemButtonPress } from 'src/ipc/messages';
+import { RendererMessenger } from 'src/ipc/RenderMessenger';
 import { IconSet } from 'widgets/Icons';
 import { useStore } from '../contexts/StoreContext';
 import { IS_MAC } from 'common/process';
